net: mptcp: implement dialMPTCP
This function is called when the user has requested MPTCP via
UseMultipathTCP() in the Dialer.

This new function fallsback to dialTCP() on operating systems that do
not support MPTCP or if MPTCP is not supported.

Note that this new function returns a TCPConn object. It could be
interesting to return a new MPTCPConn object using this composition to
have a clear distinction and maybe allow later more specific MPTCP
features without modifying the TCPConn object or its interface:

    type MPTCPConn struct {
        *TCPConn
    }

But the drawback is that if MPTCP is used by default one day (see #56539
issue on GitHub), Dial() will return a different object: MPTCPConn
instead of the previously expected TCPConn. This can probably cause
issues for apps checking the returned object.

To be discussed: maybe returning a new MPTCPConn object is not an issue
because before this patch, it was clear Dial() would return either
TCPConn or UDPConn depending on the given parameters ("tcp" or "udp").
Thus there was no need for an app to check the returned object and then
it would be OK to returned a different one, no?

Link: https://github.com/golang/go/issues/56539
Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>